Fix spurious log axes (#1555)
Issue #1554 

PR #1553 restricted options passed from Plot to Graphics to only include
valid Graphics options in order make the tests more robust. This
required adding LogPlot to the list of allowable Graphics options, with
a default of False. So now all Graphics objects have a LogPlot option,
most with a value of False, whereas previously only log plots generated
a LogPlot option in Graphics, with a value of True. However the SVG
generation code in builtin/box/graphics.py did not check the value of
the LogPlot option when deciding whether to generate log axes.