More custom get_float_value routines... (#1561)
Add more custom `get_float_value()` methods in places where it makes
sense.

Order more methods in `mathics.core.atoms.numerics` alphabetically.

@mmatera Background: For a while, I've not been happy about the way
mathics.core classes at the lowest level are defined. In particular, the
BaseElement class assumes that atoms have numeric properties and can ask
about integerness or floatness, which seems wrong. Symbol is a kind of
atom, but it is not in `mathics.core.atoms` and my simple attempts to
put it there are not drastic enough.

So I backed off all of this and focused on a smaller kind of thing,
which is to go over `get_float_values()` which is a misnomer because it
can return a complex value.