Fix format numbers (#1580)
This is another step in making the formatting more compatible with WMA.
This time, in adjusting how numbers are formatted in different forms and
representations.

I tried to stick to the form in which WMA does it, with a detail: Except
in `OutputForm`, PrecisionReal numbers in WMA are shown with all the
decimal places included in the internal representation. For example,
```
In[1]:= N[Pi,4]//InputForm                                                      

Out[1]//InputForm= 3.1415926535897932385`4.
```
In Mathics, because of our internal representation, ```N[Pi,4]```
results in a number with exactly 5 decimal digits (`3.1416`4`)

Apart from this difference, the references for the tests were collected
from WMA.

In order to make this work, some bugs and misimplementations were fixed:
* `RowBox.elements` now stores a `ListExpression` instead of the
`RobBox.items` directly.
* `String.make_boxes` uses the internal value `String._value` instead of
the proxy `String.value`.
* A more sophisticated analysis is used to determine when parenthesize
the index in a `SuperscriptBox` in its `text` representation.