Wma format comparison (#1582)
This PR adds a WL script to do the same tests that `test_format.py`
does. WMA 12 does not support `YAML` files, so I had to translate the
reference file to json. The output is also included.

To run the tests in WMA, we can open `test.m` in the Notebook interface,
or run in the command line

```
wolframscript -f test.m
```
or
```
cat test.m | math
```